feat!: Remove RecommendationsPanel (#437)
