--- src/widgets/RecommendationsPanel/api.js
@@ -1,12 +1,0 @@
-import { StrictDict } from 'utils';
-import { get, stringifyUrl } from 'data/services/lms/utils';
-import urls from 'data/services/lms/urls';
-
-export const getFetchUrl = () => (`${urls.getApiUrl()}/edx_recommendations/learner_dashboard/course_recommendations/`);
-export const apiKeys = StrictDict({ user: 'user' });
-
-const fetchRecommendedCourses = () => get(stringifyUrl(getFetchUrl()));
-
-export default {
-  fetchRecommendedCourses,
-};
