--- src/widgets/RecommendationsPanel/mockData.js
@@ -1,31 +1,0 @@
-export const recommendedCoursesData = {
-  courses: [
-    {
-      logoImageUrl: 'https://prod-discovery.edx-cdn.org/media/programs/banner_images/9a310b98-8f27-439e-be85-12d6460245c9-f2efca129273.small.jpg',
-      title: 'Recommended course 1',
-      marketingUrl: 'www.edx/recommended-course',
-      courseKey: 'cs-1',
-    },
-    {
-      logoImageUrl: 'https://prod-discovery.edx-cdn.org/media/programs/banner_images/9a310b98-8f27-439e-be85-12d6460245c9-f2efca129273.small.jpg',
-      title: 'Recommended course 2 with a really really really long name for some reason',
-      marketingUrl: 'www.edx/recommended-course',
-      courseKey: 'cs-2',
-    },
-    {
-      logoImageUrl: 'https://prod-discovery.edx-cdn.org/media/programs/banner_images/9a310b98-8f27-439e-be85-12d6460245c9-f2efca129273.small.jpg',
-      title: 'Recommended course 3',
-      marketingUrl: 'www.edx/recommended-course',
-      courseKey: 'cs-3',
-    },
-    {
-      logoImageUrl: 'https://prod-discovery.edx-cdn.org/media/programs/banner_images/9a310b98-8f27-439e-be85-12d6460245c9-f2efca129273.small.jpg',
-      title: 'Recommended course 4',
-      marketingUrl: 'www.edx/recommended-course',
-      courseKey: 'cs-4',
-    },
-  ],
-  isControl: true,
-};
-
-export default recommendedCoursesData;
